Move logic to container
